+++ .golangci.yaml
@@ -149,6 +149,8 @@ linters-settings:
         alias: testutils
       - pkg: github.com/kyma-project/telemetry-manager/internal/utils/metricpipeline
         alias: metricpipelineutils
+      - pkg: github.com/kyma-project/telemetry-manager/internal/utils/telemetry
+        alias: telemetryutils
 
   errcheck:
     check-type-assertions: true # Reports type assertions: `a := b.(SomeStruct)`.
